// src/app/Models/stop.model.ts
export class Stop {
    primarykeyduid: number;
    stopnumber: number;
    shortcode: string;
    langbezeichner: string;
    externalid?: string;
    haltestellenlangnummer: number;
    zonewabeindex: string;
    ibisbezeichner: string;
    tlpmode: string;
    fahrzeugansagetextindex: string;
    parenttableindex?: string;
    tarifortsname?: string;
    haltestellensymbolindex?: string;
    mfaanschlussbildschirm?: string;
    vorschauzeit?: string;
    importstatus: number;
    commentary?: string;
    nationalstopnumber?: string;
    stopiconforcc?: string;
    assetpictureindex?: string;
    effectonccdelta?: string;
    isdepot: string;
    longitude: number;
    latitude: number;
    haltestellenindex: string;
  }
  